# interfaces/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todo Backend API",
    "description": "REST API for managing Todo tasks (create, read, update, toggle, delete).",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Health Check",
        "description": "Health check endpoint.",
        "operationId": "health_check__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "List Todos",
        "description": "Retrieve all todos ordered by creation time descending.",
        "operationId": "get_todos_todos_get",
        "responses": {
          "200": {
            "description": "A list of todos.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TodoOut"
                  },
                  "type": "array",
                  "title": "Response Get Todos Todos Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "summary": "Create Todo",
        "description": "Create a new todo item.",
        "operationId": "create_todo_todos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Todo created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "400": {
            "description": "Validation error."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{id}": {
      "put": {
        "tags": [
          "Todos"
        ],
        "summary": "Update Todo",
        "description": "Update fields (title, description, completed) of a todo.",
        "operationId": "put_todo_todos__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "ID of the todo to update",
              "title": "Id"
            },
            "description": "ID of the todo to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated todo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "404": {
            "description": "Todo not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Todos"
        ],
        "summary": "Delete Todo",
        "description": "Delete a todo by its identifier.",
        "operationId": "delete_todo_todos__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "ID of the todo to delete",
              "title": "Id"
            },
            "description": "ID of the todo to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Todo deleted."
          },
          "404": {
            "description": "Todo not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{id}/toggle": {
      "patch": {
        "tags": [
          "Todos"
        ],
        "summary": "Toggle Todo",
        "description": "Flip the completion status of a todo.",
        "operationId": "patch_toggle_todo_todos__id__toggle_patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "ID of the todo to toggle",
              "title": "Id"
            },
            "description": "ID of the todo to toggle"
          }
        ],
        "responses": {
          "200": {
            "description": "Toggled todo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "404": {
            "description": "Todo not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "TodoCreate": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Short title of the task"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional detailed description"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TodoCreate"
      },
      "TodoOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Unique identifier"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the task"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Detailed description"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "description": "Completion status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "completed"
        ],
        "title": "TodoOut"
      },
      "TodoUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Updated title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Updated description"
          },
          "completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed",
            "description": "Updated completion flag"
          }
        },
        "type": "object",
        "title": "TodoUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "Service health and diagnostics"
    },
    {
      "name": "Todos",
      "description": "Operations for managing todo tasks"
    }
  ]
}